+++ src/invoices/InvoiceLineDetails/ReceivingHistory/useReceivingHistory.js
@@ -10,16 +10,17 @@ import {
   PIECE_STATUS,
 } from '@folio/stripes-acq-components';
 
-export const useReceivingHistory = (poLineId) => {
+export const useReceivingHistory = (poLineId, { pagination }) => {
   const ky = useOkapiKy();
   const [namespace] = useNamespace({ key: 'invoice-line-receiving-history' });
 
   const { isLoading, data = {} } = useQuery(
-    [namespace, poLineId],
+    [namespace, poLineId, pagination],
     () => ky.get(ORDER_PIECES_API, {
       searchParams: {
         query: `poLineId==${poLineId} and receivingStatus==${PIECE_STATUS.received} sortby receivedDate/sort.descending`,
-        limit: LIMIT_MAX,
+        limit: pagination.limit ?? LIMIT_MAX,
+        offset: pagination.offset ?? 0,
       },
     }).json(),
     { enabled: Boolean(poLineId) },
@@ -27,6 +28,7 @@ export const useReceivingHistory = (poLineId) => {
 
   return {
     pieces: data.pieces || [],
+    piecesCount: data.totalRecords,
     isLoading,
   };
 };