+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * 'Undefined' displayed for receipt status and payment status if Invoice Line was created without POL. Refs UINV-350.
 * Adding tenant's timezone in UI. Refs UINV-351.
 * Settings > Invoices > change focus. Refs UINV-344.
+* Add pagination to Receiving history accordion MCL table. Refs UINV-353.
 
 ## [3.0.2](https://github.com/folio-org/ui-invoice/tree/v3.0.2) (2021-11-25)
 [Full Changelog](https://github.com/folio-org/ui-invoice/compare/v3.0.1...v3.0.2)